requirements.txt
langchain
langchain-openai
langchain-google-genai
langchain-tavily
python-dotenv